Normative: Get receiver fields from a PlainDateTime in ZonedDateTime.with()
Instead of calling PrepareTemporalFields on the receiver in
ZonedDateTime.p.with() to get the values of the date units, first create
a PlainDateTime from the ZonedDateTime's exact time and time zone, and
call PrepareTemporalFields on that.

This still calls the corresponding calendar method for each field, but
avoids calling the time zone's getOffsetNanosecondsFor() method
redundantly for each field.