Normative: Look up time zone methods only once
Introduce a new spec type, Time Zone Record, which stores a time zone
object's getOffsetNanosecondsFor and getPossibleInstantsFor methods once
they have been observably looked up. The record is passed around into
abstract operations instead of the time zone object itself

Most API operations that do time zone calculations need to look up both
getOffsetNanosecondsFor and getPossibleInstantsFor unconditionally. Only
in the case where we convert a plain time to an exact time, it's possible
we only need to look up getPossibleInstantsFor, or neither (if the offset
is supplied.)

This is a large commit but most of it is mechanical replacement of
Temporal.TimeZone variables with Time Zone Record variables.