Normative: Fast-path differences between identical objects
If two Temporal objects have identical internal slots, then we can skip
calculating the difference with a calendar method; the difference is
always 0.

(This optimization isn't necessary for PlainTime or Instant differences,
since no calendar methods are called for those.)

Note that the early return still comes after the processing of the options
object. Passing an illegal options object to until() or since() should
still throw, even if the difference is zero.