Normative: Remove loops in RoundDuration
In RoundDuration, we already calculated years rounding without a loop. I'm
not sure why we didn't do this for months and weeks, but it's not
necessary to use a loop for those either.

This should not affect any results, but has an observable effect on calls
to custom calendar methods.